requirements.txt
accelerate==0.31.0
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
alabaster==0.7.16
albumentations==1.4.3
app-model==0.2.7
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
Babel==2.15.0
build==1.2.1
cachey==0.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
clip==1.0
clip-interrogator==0.6.0
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
controlnet_aux==0.0.7
cycler==0.12.1
dask==2024.6.2
dask-expr==1.1.6
dask-image==2024.5.3
datasets==3.0.1
debugpy==1.8.2
decorator==5.1.1
deeplabcut==3.0.0rc1
diffusers==0.30.3
dill==0.3.8
dlclibrary==0.0.6
docker-pycreds==0.4.0
docstring_parser==0.16
docutils==0.17.1
einops==0.8.0
executing==2.0.1
filelock==3.15.4
filterpy==1.4.5
flash_attn==2.7.4.post1
flexcache==0.3
flexparser==0.3.1
fonttools==4.53.0
freetype-py==2.4.0
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
glide-text2im==0.0.0
HeapDict==1.0.1
hsluv==5.0.4
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
imagesize==1.4.1
imgaug==0.4.0
importlib_metadata==8.0.0
in-n-out==0.2.1
ipykernel==6.29.4
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
llava==1.7.0.dev0
llvmlite==0.43.0
locket==1.0.0
magicgui==0.8.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
napari==0.4.18
napari-console==0.0.9
napari-deeplabcut==0.2.1.7
napari-plugin-engine==0.2.0
napari-svg==0.2.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.3
npe2==0.7.6
numba==0.60.0
numexpr==2.10.0
numpy==1.26.4
numpydoc==1.5.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.26.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
partd==1.4.2
patsy==0.5.6
peft==0.11.1
pexpect==4.9.0
pillow==10.3.0
PIMS==0.7
Pint==0.24.1
pip==24.0
platformdirs==4.2.2
pooch==1.8.2
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
psygnal==0.11.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pycocotools==2.0.8
pyconify==0.1.6
pydantic==1.10.17
pydantic-compat==0.1.2
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.1.2
pyproject_hooks==1.1.0
PySide6==6.4.2
PySide6-Addons==6.4.2
PySide6-Essentials==6.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
QDarkStyle==3.1
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.10.1
segment-anything==1.0
segment-anything-hq==0.3
sentence-transformers==3.2.0
sentencepiece==0.2.0
sentry-sdk==2.7.1
setproctitle==1.3.3
setuptools==69.5.1
shapely==2.0.4
shellingham==1.5.4
shiboken6==6.4.2
six==1.16.0
slicerator==1.1.0
smmap==5.0.1
snowballstemmer==2.2.0
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.3
statsmodels==0.14.2
superqt==0.6.7
sympy==1.12.1
tables==3.8.0
threadpoolctl==3.5.0
tifffile==2024.6.18
timm==1.0.7
tokenizers==0.20.0
tomli_w==1.0.0
toolz==0.12.1
torch==2.4.0
torchtyping==0.1.5
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.45.2
trimesh==4.4.2
triton==3.0.0
typeguard==2.13.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vispy==0.12.2
wandb==0.17.3
wcwidth==0.2.13
wheel==0.43.0
wrapt==1.16.0
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.13.1
zipp==3.19.2
